move route_provider to separate mod
