+++ src/routing/ic/mod.rs
@@ -1,6 +1,6 @@
 pub mod health_check;
+pub mod route_provider;
 pub mod transport;
-
 use std::sync::Arc;
 
 use anyhow::Error;
